Moving the AppInsight under context
