+++ pages/_app.tsx
@@ -15,7 +15,7 @@ import ZendeskButton from "../components/zendeskButton/zendeskButton";
 const zendesk = process.env.NEXT_PUBLIC_ZENDESK_CHAT_KEY;
 // Hack as per https://stackoverflow.com/a/66575373 to stop font awesome icons breaking
 import "@fortawesome/fontawesome-svg-core/styles.css";
-import AzureAppInsights from "../services/appInsight_ui";
+import AzureAppInsights from "../context/appInsight_ui";
 
 const isDev = process.env.NODE_ENV === "development";
